+++ contracts/mocks/token/ERC1155/REVVMotorsportVouchersMock.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.15;
+pragma solidity 0.8.17;
 
 import {IForwarderRegistry} from "@animoca/ethereum-contracts/contracts/metatx/interfaces/IForwarderRegistry.sol";
 import {REVVMotorsportVouchers} from "./../../../token/ERC1155/REVVMotorsportVouchers.sol";